VantageCloud_Lake/UseCases/Medical_Visual_Question_Answering_BYOLLM/requirements.txt
adjustText==0.8
aia==0.2.0
alembic==1.7.7
altair==4.2.0
annotated-types==0.7.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
async-generator==1.10
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
bokeh==2.4.2
Bottleneck==1.3.4
branca==0.8.0
brotlipy==0.7.0
cached-property==1.5.2
certifi==2024.8.30
certipy==0.1.3
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.0.0
cmake==3.30.5
cmdstanpy==1.2.4
colorama==0.4.4
coloredlogs==15.0.1
colorlover==0.3.0
comm==0.2.3
zipp==3.8.0
conda==4.12.0
conda-package-handling==1.8.1
contourpy==1.3.0
convertdate==2.4.0
cryptography==43.0.3
cycler==0.11.0
Cython==0.29.28
cytoolz==0.11.2
dask==2022.4.0
debugpy==1.5.1
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.4
distributed==2022.4.0
docker==5.0.3
entrypoints==0.4
ephem==4.1.6
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.1
fastjsonschema==2.15.3
filelock==3.16.1
fiona==1.10.1
fire==0.7.0
flatbuffers==24.3.25
flit_core==3.7.1
folium==0.14.0
fonttools==4.31.2
fsspec==2024.10.0
future==1.0.0
geopandas==0.12.2
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.1.2
greenlet==1.1.2
h2o==3.40.0.4
h5py==3.6.0
HeapDict==1.0.1
hf-xet==1.1.10
holidays==0.59
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.3
imagecodecs==2022.2.22
imageio==2.16.1
imbalanced-learn==0.12.4
imblearn==0.0
iminuit==2.21.0
importlib-metadata==4.11.3
importlib-resources==5.6.0
install-jdk==1.0.4
ipykernel==6.12.0
ipympl==0.8.8
ipython==8.18.1
ipython-genutils==0.2.0
ipython-sql==0.4.1
ipywidgets==8.1.7
itables==2.5.2
jdk4py==17.0.3.0
jedi==0.19.2
Jinja2==3.1.1
joblib==1.4.2
json5==0.9.5
jsonschema==4.4.0
jupyter-client==7.2.1
jupyter-core==4.9.2
jupyter-server==1.16.0
jupyter-telemetry==0.1.0
jupyterhub==2.2.2
jupyterlab==3.3.2
jupyterlab-google-analytics==0.2.0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.12.0
jupyterlab_widgets==3.0.15
keras2onnx==1.7.0
kiwisolver==1.4.2
libmambapy==0.22.1
lightgbm==4.6.0
lit==18.1.8
llvmlite==0.38.0
locket==0.2.0
LunarCalendar==0.0.9
lxml==5.3.0
lz4==4.0.0
Mako==1.2.0
mamba==0.22.1
MarkupSafe==2.1.1
matplotlib==3.9.4
matplotlib-inline==0.1.7
mistune==0.8.4
mpmath==1.2.1
msgpack==1.0.3
munkres==1.1.4
nbclassic==0.3.7
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.3.0
nest-asyncio==1.5.5
networkx==2.7.1
notebook==6.4.10
notebook-shim==0.1.0
numba==0.55.1
numexpr==2.8.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.8.90
nyoka==4.3.0
oauthlib==3.2.2
onnx==1.13.1
onnxconverter-common==1.7.0
onnxmltools==1.7.0
onnxruntime==1.19.2
onnxruntime-tools==1.7.0
opencv-python==4.10.0.84
openpyxl==3.1.1
packaging==21.3
pamela==1.0.0
pandas==2.1.3
pandocfilters==1.5.0
parso==0.8.5
partd==1.2.0
patsy==0.5.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
pip==22.0.4
plotly==5.8.2
prettytable==0.7.2
prometheus-client==0.13.1
prompt_toolkit==3.0.52
prophet==1.1.4
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py3nvml==0.2.7
pycosat==0.6.3
pycparser==2.21
pycryptodome==3.14.1
pycurl==7.45.1
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.9.0
PyMeeus==0.5.12
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyproj==3.6.1
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.1.1
python-json-logger==2.0.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0.3
pyzmq==22.3.0
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rpy2==3.4.5
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
ruamel-yaml-conda==0.15.80
safetensors==0.6.2
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.11.2
seaborn==0.13.2
Send2Trash==1.8.0
sentence-transformers==5.1.1
sentencepiece==0.2.1
setuptools==62.0.0
shapely==2.0.6
simplegeneric==0.8.1
six==1.16.0
skl2onnx==1.11.2
sklearn-pandas==2.2.0
sklearn2pmml==0.91.0
smmap==5.0.1
sniffio==1.2.0
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.3.1
SQLAlchemy==2.0.23
sqlparse==0.5.1
stack-data==0.6.3
stanio==0.5.1
statsmodels==0.13.2
sympy==1.14.0
tables==3.7.0
tabulate==0.9.0
tblib==1.7.0
tdnpathviz==0.1.2.22
tdsense==0.1.3.11
tenacity==8.0.1
teradata-connection-manager-prebuilt==3.3.0
teradata-database-explorer-prebuilt==3.3.0
teradata-preferences-prebuilt==3.3.0
teradata-resultset-renderer-prebuilt==3.3.0
teradata-sqlhighlighter-prebuilt==3.3.0
teradatagenai==20.0.0.2
teradataml==20.0.0.7
teradatamlwidgets==20.0.0.6
teradatamodelops==7.0.3
teradatasql==20.0.0.34
teradatasqlalchemy==20.0.0.7
termcolor==2.5.0
terminado==0.13.3
testpath==0.6.0
tf2onnx==1.16.1
threadpoolctl==3.1.0
tifffile==2022.3.25
tokenizers==0.22.1
toolz==0.11.2
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tornado==6.1
tqdm==4.64.0
traitlets==5.14.3
transformers==4.56.2
triton==3.4.0
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2022.1
tzlocal==4.2
unicodedata2==14.0.0
urllib3==1.26.9
wcwidth==0.2.14
webencodings==0.5.1
websocket-client==1.3.2
wheel==0.37.1
widgetsnbextension==4.0.14
wordcloud==1.9.4
xgboost==1.7.3
xlrd==2.0.1
xmltodict==0.14.2
zict==2.1.0